requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gunicorn==21.2.0
piper-tts==1.2.0
huggingface_hub